Added half-res mode, lz4 compression, started converting all UI to 16bpp HDLIB is now a linked asm file, not a standalone library. Added err enum for future handling guiUtils is now header only file Temporarily disabled screen shifting
